pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neoskidrl"
version = "0.1.0"
description = "NeoSkidRL: skid-steer navigation with MuJoCo + Gymnasium"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.23",
  "pyyaml>=6.0",
  "gymnasium>=0.29",
  "mujoco>=3.1.0",
]

[project.optional-dependencies]
train = [
  "stable-baselines3>=2.3.0",
  "tensorboard>=2.13",
  "tqdm>=4.66",
]
viz = [
  "matplotlib>=3.7",
]
test = [
  "pytest>=7.4",
]
video = [
  "imageio[ffmpeg]>=2.31",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
neoskidrl = ["models/*.xml", "config/*.yml"]
